feat(modules)!: Add BalancedImageFolder class for oversampling and undersampling
This commit implements the BalancedImageFolder class in a new module file, and it has been incorporated into the main training script. The additions made provide a clean and customizable approach to the problem of class imbalance in datasets.

BREAKING CHANGE: Add BalancedImageFolder class for oversampling and undersampling

This new functionality supersedes the previous undersampling and oversampling methods in the train.py script, which were removed.